server/payment_service/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
confluent-kafka==2.5.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
stripe==10.8.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
